Changing materials to use better parameters
