+++ Echo.Core/src/Evaluation/Materials/Conductor.cs
@@ -1,7 +1,5 @@
-﻿using System;
-using Echo.Core.Aggregation.Primitives;
+﻿using Echo.Core.Aggregation.Primitives;
 using Echo.Core.Common.Diagnostics;
-using Echo.Core.Common.Mathematics;
 using Echo.Core.Common.Memory;
 using Echo.Core.Evaluation.Scattering;
 using Echo.Core.Textures;
@@ -9,7 +7,7 @@ using Echo.Core.Textures.Colors;
 
 namespace Echo.Core.Evaluation.Materials;
 
-public class Glass : Material
+public class Conductor : Material
 {
 	NotNull<Texture> _roughness = Pure.black;
 	NotNull<Texture> _refractiveIndex = Pure.white;
@@ -33,15 +31,20 @@ public class Glass : Material
 		var albedo = (RGB128)SampleAlbedo(contact);
 		if (albedo.IsZero) return;
 
-		float index = Sample(RefractiveIndex, contact).R;
+		RGB128 index = Sample(RefractiveIndex, contact);
 		RGB128 roughness = Sample(Roughness, contact);
-		float roughnessU = FastMath.Clamp01(roughness.R);
-		float roughnessV = FastMath.Clamp01(roughness.G);
 
-		if (!FastMath.AlmostZero(roughnessU) || !FastMath.AlmostZero(roughnessV))
-		{
-			throw new NotImplementedException();
-		}
-		else make.Add<SpecularFresnel>().Reset(albedo, 1f, index);
+		// float alphaX = IMicrofacet.GetAlpha(FastMath.Clamp01(roughness.R));
+		// float alphaY = IMicrofacet.GetAlpha(FastMath.Clamp01(roughness.G));
+
+		float alphaX = roughness.R;
+		float alphaY = roughness.G;
+
+		make.Add<GlossyReflection<TrowbridgeReitzMicrofacet, ConductorFresnel>>().Reset
+		(
+			RGB128.White,
+			new TrowbridgeReitzMicrofacet(alphaX, alphaY),
+			new ConductorFresnel(RGB128.White, index, albedo)
+		);
 	}
 }
